refactor(extensions-hydration): split useExtensionsHydration into per-concern submodules
The hook owns extension classification, sidebar row generation, theme
injection, layout activation, keybinding/event-route/frontend-module
hydration, and slash command merging — broad enough to be hard to
reason about in one file. Split into a `src/hooks/extensionsHydration/`
directory with one submodule per concern and a thin facade that
preserves the public surface.

Submodule layout:
- types: all exported types + BUILTIN_THEMES
- classification: project/user scope helpers and filters
- sidebarRows: buildExtensionSidebarItems + useHydrateExtensions
- themes: injectThemeStyle, useThemeActions
- layouts: summarizeLayoutComponents, useLayoutActions
- keybindings, eventRoutes, frontendModules, slashCommands: one per
  hydrate channel

Factory hooks use the `useX` naming required by react-hooks/refs since
they receive ref objects from the parent hook. Tests move next to their
implementation modules (classification.test.ts + sidebarRows.test.ts).

Behavior unchanged; every public name remains importable from
`./useExtensionsHydration`.
