+++ src/hooks/extensionsHydration/sidebarRows.test.ts
@@ -1,65 +1,5 @@
 import { describe, expect, it } from "vitest";
-import {
-  buildExtensionSidebarItems,
-  filterExtensionSummariesByProject,
-} from "./useExtensionsHydration";
-
-describe("filterExtensionSummariesByProject", () => {
-  it("keeps global extensions and only the active project's local extensions", () => {
-    const filtered = filterExtensionSummariesByProject(
-      [
-        { name: "user-ext", source: "directory" },
-        {
-          name: "mold:image-gallery",
-          source: "project-directory",
-          projectRoot: "/repo/mold",
-        },
-        {
-          name: "latent:tools",
-          source: "project-directory",
-          projectRoot: "/repo/latentforge",
-        },
-      ],
-      "/repo/latentforge",
-    );
-
-    expect(filtered.map((e) => e.name)).toEqual([
-      "user-ext",
-      "latent:tools",
-    ]);
-  });
-
-  it("drops project extensions when no active project path is known", () => {
-    const filtered = filterExtensionSummariesByProject(
-      [
-        { name: "user-ext", source: "directory" },
-        {
-          name: "mold:image-gallery",
-          source: "project-directory",
-          projectRoot: "/repo/mold",
-        },
-      ],
-      null,
-    );
-
-    expect(filtered.map((e) => e.name)).toEqual(["user-ext"]);
-  });
-
-  it("does not treat sibling path prefixes as the same project", () => {
-    const filtered = filterExtensionSummariesByProject(
-      [
-        {
-          name: "mold:image-gallery",
-          source: "project-directory",
-          projectRoot: "/repo/mold",
-        },
-      ],
-      "/repo/mold-tools",
-    );
-
-    expect(filtered).toEqual([]);
-  });
-});
+import { buildExtensionSidebarItems } from "./sidebarRows";
 
 describe("buildExtensionSidebarItems", () => {
   it("surfaces user, project, failed, and disabled extensions without the core layout", () => {
@@ -401,7 +341,9 @@ describe("buildExtensionSidebarItems", () => {
     // `@mold/...` folds into project (scope matches active project);
     // `@brink/...` stays user (unrelated scope). Sort order: project
     // bucket first, alphabetical within.
-    expect(items.map((item) => ({ label: item.label, hint: item.hint }))).toEqual([
+    expect(
+      items.map((item) => ({ label: item.label, hint: item.hint })),
+    ).toEqual([
       { label: "@mold/image-gallery-ui", hint: "project · disabled" },
       { label: "mold:image-gallery", hint: "project · disabled" },
       { label: "@brink/current-context-widget", hint: "user · disabled" },
